# specs/001-sitemap-monitor/contracts/api.yaml
openapi: 3.1.0
info:
  title: Sitemap Monitor API
  description: API for monitoring sitemap changes
  version: 1.0.0
  contact:
    name: Sitemap Monitor Team
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /users/me:
    get:
      tags: [users]
      summary: 获取当前用户信息
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    patch:
      tags: [users]
      summary: 更新当前用户信息
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /users/me/password:
    put:
      tags: [users]
      summary: 修改密码
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: 密码修改成功
        '400':
          description: 当前密码错误
  /users/me/onboarding:
    post:
      tags: [users]
      summary: 完成新手引导
      operationId: completeOnboarding
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 引导完成

  # ============ 监控任务 ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        is_verified:
          type: boolean
        has_completed_onboarding:
          type: boolean
        created_at:
          type: string
          format: date-time
...